+++ TapSDK/TapBootstrapSDK/IStartTask.cs
@@ -1,0 +1,7 @@
+namespace TapTap.Bootstrap
+{
+    public interface IStartTask
+    {
+        void Invoke(TapConfig config);
+    }
+}