+++ TapSDK/TapBootstrapSDK/LoginType.cs
@@ -1,7 +1,0 @@
-namespace TapTap.Bootstrap
-{
-    public enum LoginType
-    {
-        TAPTAP = 0
-    }
-}