+++ TapSDK/TapBootstrapSDK/TapBootstrap.cs
@@ -1,5 +1,3 @@
-using System;
-
 namespace TapTap.Bootstrap
 {
     public class TapBootstrap
@@ -7,63 +5,6 @@ namespace TapTap.Bootstrap
         public static void Init(TapConfig tapConfig)
         {
             TapBootstrapImpl.GetInstance().Init(tapConfig);
-        }
-
-        public static void Login(LoginType loginType, string[] permissions)
-        {
-            TapBootstrapImpl.GetInstance().Login(loginType, permissions);
-        }
-
-        [Obsolete("已废弃")]
-        public static void Bind(LoginType loginType, string[] permissions)
-        {
-            TapBootstrapImpl.GetInstance().Bind(loginType, permissions);
-        }
-
-        public static void GetTestQualification(Action<bool, TapError> action)
-        {
-            TapBootstrapImpl.GetInstance().GetTestQualification(action);
-        }
-
-        public static void RegisterUserStatusChangedListener(ITapUserStatusChangedListener listener)
-        {
-            TapBootstrapImpl.GetInstance().RegisterUserStatusChangedListener(listener);
-        }
-
-        public static void RegisterLoginResultListener(ITapLoginResultListener listener)
-        {
-            TapBootstrapImpl.GetInstance().RegisterLoginResultListener(listener);
-        }
-
-        public static void GetUser(Action<TapUser, TapError> action)
-        {
-            TapBootstrapImpl.GetInstance().GetUser(action);
-        }
-
-        public static void GetDetailUser(Action<TapUserDetail, TapError> action)
-        {
-            TapBootstrapImpl.GetInstance().GetDetailUser(action);
-        }
-
-        public static void GetAccessToken(Action<AccessToken, TapError> action)
-        {
-            TapBootstrapImpl.GetInstance().GetAccessToken(action);
-        }
-
-        public static void Logout()
-        {
-            TapBootstrapImpl.GetInstance().Logout();
-        }
-        
-        [Obsolete("已弃用")]
-        private static void OpenUserCenter()
-        {
-            // TapBootstrapImpl.GetInstance().OpenUserCenter();
-        }
-
-        public static void SetPreferLanguage(TapLanguage tapLanguage)
-        {
-            TapBootstrapImpl.GetInstance().SetPreferLanguage(tapLanguage);
         }
     }
 }