+++ TapSDK/TapBootstrapSDK/TapStorageStartTask.cs
@@ -1,0 +1,14 @@
+using LeanCloud;
+using LeanCloud.Storage;
+
+namespace TapTap.Bootstrap
+{
+    public class TapStorageStartTask : IStartTask
+    {
+        public void Invoke(TapConfig config)
+        {
+            LCApplication.Initialize(config.ClientID, config.ClientToken, config.ServerURL);
+            LCObject.RegisterSubclass("_User", () => new TDSUser());
+        }
+    }
+}