+++ TapSDK/TapBootstrapSDK/TapUserStatusChangedListener.cs
@@ -1,9 +1,0 @@
-namespace TapTap.Bootstrap
-{
-    public interface ITapUserStatusChangedListener
-    {
-        void OnLogout(TapError error);
-
-        void OnBind(TapError error);
-    }
-}