+++ Assets/TapTap/Common/CHANGELOG.md
@@ -1,10 +1,20 @@
 # ChangeLog
 
+## 3.0.0
+
+### Optimization
+
+- Bridge 新增 Task 桥接
+
+## 2.1.7
+
+### None
+
 ## 2.1.6
 
 ### BreakingChanges
 
-- 修改 **TapCommon.OpenReviewInTapGlobal()** 方法命名 
+- 修改 **TapCommon.OpenReviewInTapGlobal()** 方法命名
 
 ## 2.1.5