+++ Assets/TapTap/Login/CHANGELOG.md
@@ -1,5 +1,34 @@
 # ChangeLog
 
+## 3.0.0
+
+### New Feature
+
+- 新增 TapTap OAuth 相关接口
+  ```
+  // 登陆  
+  TapLogin.Login();
+  // 登出
+  TapLogin.Logout();
+  ```
+- 新增篝火测试资格
+  ```
+  var boolean = await TapLogin.GetTestQualification();
+  ```
+### Dependencies
+
+- TapTap.Common v3.0.0
+
+## 2.1.7
+
+### Optimization
+
+- 更新 iOS 拦截 openUrl 方式
+
+### Dependencies
+
+- TapTap.Common v2.1.7
+
 ### 2.1.6
 
 ### Optimization and fixed bugs